+++ examples/Mythology/API.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric  #-}
+{-# LANGUAGE RankNTypes     #-}
 {-# LANGUAGE TypeOperators  #-}
 
 module Mythology.API
@@ -10,7 +11,7 @@ import qualified Data.ByteString.Lazy.Char8 as B
 
 import           Data.Morpheus              (interpreter)
 import           Data.Morpheus.Kind         (GQLArgs, GQLQuery)
-import           Data.Morpheus.Types        ((::->) (..), GQLRoot (..))
+import           Data.Morpheus.Types        ((::->), GQLRoot (..), Resolver (..))
 import           Data.Text                  (Text)
 import           GHC.Generics               (Generic)
 import           Mythology.Character.Deity  (Deity (..), dbDeity)