+++ src/Data/Morpheus/Kind/InputRouter.hs
@@ -10,23 +10,18 @@
 module Data.Morpheus.Kind.InputRouter where
 
 import           Data.Morpheus.Error.Internal       (internalArgumentError, internalTypeMismatch)
+import           Data.Morpheus.Generics.EnumRep     (EnumRep (..))
 import           Data.Morpheus.Generics.GDecode     (GDecode (..))
-import qualified Data.Morpheus.Kind.GQLEnum         as E (EnumConstraint, decode, introspect)
-import qualified Data.Morpheus.Kind.GQLInputObject  as I (IObjectConstraint, decode, inputField, introspect)
+import           Data.Morpheus.Kind.GQLKinds        (Decode_, EnumConstraint, IField_, IObjectConstraint, Intro_,
+                                                     introspectEnum, introspectInputObject)
 import qualified Data.Morpheus.Kind.GQLScalar       as S (GQLScalar (..))
 import           Data.Morpheus.Kind.GQLType         (GQLType, field_)
-import           Data.Morpheus.Kind.Internal        (Decode_, ENUM, IField_, INPUT_OBJECT, Intro_, KIND, SCALAR,
-                                                     WRAPPER)
+import           Data.Morpheus.Kind.Internal        (ENUM, INPUT_OBJECT, KIND, SCALAR, WRAPPER)
 import           Data.Morpheus.Kind.Utils           (listField, maybeField)
 import           Data.Morpheus.Schema.TypeKind      (TypeKind (..))
 import           Data.Morpheus.Types.Internal.Value (Value (..))
 import           Data.Proxy                         (Proxy (..))
 import           GHC.Generics
-
-class InputTypeRouter a b where
-  __introspect :: Proxy b -> Intro_ a
-  __decode :: Proxy b -> Decode_ a
-  __field :: Proxy b -> IField_ a
 
 _field ::
      forall a. InputTypeRouter a (KIND a)
@@ -43,6 +38,11 @@ _introspect ::
   => Intro_ a
 _introspect = __introspect (Proxy @(KIND a))
 
+class InputTypeRouter a b where
+  __introspect :: Proxy b -> Intro_ a
+  __decode :: Proxy b -> Decode_ a
+  __field :: Proxy b -> IField_ a
+
 instance (InputTypeRouter a (KIND a)) => GDecode Value (K1 i a) where
   gDecode key' (Object object) =
     case lookup key' object of
@@ -55,16 +55,17 @@ instance (S.GQLScalar a, GQLType a) => InputTypeRouter a SCALAR where
   __introspect _ _ = S.introspect (Proxy @a)
   __field _ _ = field_ SCALAR (Proxy @a) ()
 
-instance E.EnumConstraint a => InputTypeRouter a ENUM where
-  __decode _ (Enum value) = pure (E.decode value)
+instance EnumConstraint a => InputTypeRouter a ENUM where
+  __decode _ (Enum value) = pure (to $ gToEnum value)
   __decode _ isType       = internalTypeMismatch "Enum" isType
-  __introspect _ _ = E.introspect (Proxy @a)
   __field _ _ = field_ ENUM (Proxy @a) ()
+  __introspect _ _ = introspectEnum (Proxy @a)
 
-instance I.IObjectConstraint a => InputTypeRouter a INPUT_OBJECT where
-  __decode _ = I.decode
-  __field _ = I.inputField
-  __introspect _ = I.introspect
+instance IObjectConstraint a => InputTypeRouter a INPUT_OBJECT where
+  __decode _ (Object x) = to <$> gDecode "" (Object x)
+  __decode _ isType     = internalTypeMismatch "InputObject" isType
+  __field _ _ = field_ INPUT_OBJECT (Proxy @a) ()
+  __introspect _ = introspectInputObject
 
 instance InputTypeRouter a (KIND a) => InputTypeRouter (Maybe a) WRAPPER where
   __decode _ Null = pure Nothing