+++ src/Data/Morpheus/Kind/Utils.hs
@@ -1,15 +1,9 @@
 module Data.Morpheus.Kind.Utils
   ( maybeField
   , listField
-  , encodeMaybe
-  , encodeList
   ) where
 
-import           Data.Morpheus.Types.Internal.AST.Selection (Selection (..), SelectionRec (..))
-import           Data.Morpheus.Types.Internal.Data          (DataField (..), DataTypeWrapper (..))
-import           Data.Morpheus.Types.Internal.Validation    (ResolveIO)
-import           Data.Morpheus.Types.Internal.Value         (Value (..))
-import           Data.Text                                  (Text)
+import           Data.Morpheus.Types.Internal.Data (DataField (..), DataTypeWrapper (..))
 
 maybeField :: DataField a -> DataField a
 maybeField field@DataField {fieldTypeWrappers = NonNullType:xs} = field {fieldTypeWrappers = xs}
@@ -17,13 +11,3 @@ maybeField field                                                = field
 
 listField :: DataField a -> DataField a
 listField x = x {fieldTypeWrappers = [NonNullType, ListType] ++ fieldTypeWrappers x}
-
-type Encode a = (Text, Selection) -> a -> ResolveIO Value
-
-encodeList :: Encode a -> Encode [a]
-encodeList _ (_, Selection {selectionRec = SelectionField {}}) _ = pure $ List []
-encodeList f query list                                          = List <$> mapM (f query) list
-
-encodeMaybe :: Encode a -> Encode (Maybe a)
-encodeMaybe _ _ Nothing          = pure Null
-encodeMaybe f query (Just value) = f query value