+++ src/Data/Morpheus/Schema/Field.hs
@@ -1,10 +1,14 @@
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE TypeFamilies  #-}
 
 module Data.Morpheus.Schema.Field where
 
+import           Data.Morpheus.Kind.Internal     (KIND, OBJECT)
 import           Data.Morpheus.Schema.InputValue (InputValue)
 import           Data.Text                       (Text)
 import           GHC.Generics
+
+type instance KIND (Field a) = OBJECT
 
 data Field t = Field
   { name              :: Text